fix(backend): 图片记账的 SQL 类型冲突，导致图片体积从未计入配额
与上一个提交同一类错误的第二处，由用户报「上传了图但用量只显示 9.4k」发现。

原因：bytes 列是 bigint，INSERT 的目标列把 $3 推成 bigint；而 WHERE 里
SUM(bytes) 返回 numeric（Postgres 对 bigint 求和会提升类型防溢出），
`+ $3` 那处把 $3 推成 numeric。同一参数两种类型 → SQLSTATE 42P08，
整条语句在 Postgres 侧 prepare 就失败。修法是 $3::bigint。

后果：这个功能从来没工作过。image_objects 全库 0 条，图片体积完全不计入配额，
用户看到的「云端存储用量」只有文档正文。实测某账号：文档 9,553 字节，
而 21 张图实际 5,363,750 字节 —— 少算了 99.8%。

为什么完全无感：Worker 侧对记账失败的处理是「放行上传，只记一行日志」
（images.ts 的 recordUsage —— 宁可少算也不让用户贴不了图）。这个取舍本身对，
但它让「永久失败」也一样静默，唯一痕迹是 Worker 日志里的 warn。

prepare 测试改成自动枚举本包所有非测试 .go，不再手写文件名列表。
上一个提交里我手写了四个文件名，正好漏掉 image_quota.go —— 列表要靠人维护
就一定会漏，这次是现成的证据。现在覆盖 44 条 SQL、10 个文件。

变异验证：两处 ::bigint 都去掉（真正的 bug 现场）会被抓住。只去掉一处不会 ——
剩下那处仍钉住类型，代码依然正确，不是漏报。

+++ backend/internal/server/image_quota.go
@@ -99,15 +99,30 @@ func (a *App) recordImageObject(
 	//
 	// 判定里要把文档正文也算上 —— 配额是"云端存储"而不是"图床"，
 	// 少算文档会让一个写了几百篇长文的人凭空多出配额
+	// $3 必须显式标注 ::bigint，否则整条语句在 Postgres 侧 prepare 就失败。
+	//
+	// 原因：bytes 列是 bigint，所以 INSERT 的目标列把 $3 推成 bigint；而下面
+	// WHERE 里 SUM(bytes) 返回的是 numeric（Postgres 对 bigint 求和会提升类型，
+	// 防溢出），于是 `+ $3` 那处把 $3 推成 numeric。同一参数两种类型 →
+	// 「inconsistent types deduced for parameter $3」（SQLSTATE 42P08）。
+	//
+	// 后果是这个功能**从来没工作过**：记账每次都 500，image_objects 全库 0 条，
+	// 于是图片体积完全不计入配额 —— 用户看到的"云端存储用量"只有文档正文。
+	// 而 Worker 侧对记账失败的处理是「放行上传，只记一行日志」（见 images.ts 的
+	// recordUsage：宁可少算也不让用户贴不了图），所以前端完全无感，
+	// 唯一的痕迹是 Worker 日志里的一行 warn。
+	//
+	// 与 documents.go 里新建文档那条是同一类错误（同一天发现的两处），
+	// 都由 TestDocumentSQLPrepares 之外的 TestImageQuotaSQLPrepares 兜住。
 	tag, err := a.db.Exec(ctx, `
 		INSERT INTO image_objects (object_key, user_id, bytes)
-		SELECT $1, $2, $3
+		SELECT $1, $2, $3::bigint
 		WHERE COALESCE(
 			(SELECT SUM(bytes) FROM image_objects WHERE user_id = $2), 0
 		) + COALESCE(
 			(SELECT SUM(octet_length(content) + octet_length(title))
 			 FROM documents WHERE user_id = $2), 0
-		) + $3 <= $4
+		) + $3::bigint <= $4
 		ON CONFLICT (object_key) DO NOTHING
 	`, key, userID, bytes, a.imageQuota())
 	if err != nil {
